Changed the filestructure to follow the labcode, started implementing the node structure
