--- labCode/kademlia.go
@@ -1,16 +1,0 @@
-package d7024e
-
-type Kademlia struct {
-}
-
-func (kademlia *Kademlia) LookupContact(target *Contact) {
-	// TODO
-}
-
-func (kademlia *Kademlia) LookupData(hash string) {
-	// TODO
-}
-
-func (kademlia *Kademlia) Store(data []byte) {
-	// TODO
-}
