--- labCode/network.go
@@ -1,24 +1,0 @@
-package d7024e
-
-type Network struct {
-}
-
-func Listen(ip string, port int) {
-	// TODO
-}
-
-func (network *Network) SendPingMessage(contact *Contact) {
-	// TODO
-}
-
-func (network *Network) SendFindContactMessage(contact *Contact) {
-	// TODO
-}
-
-func (network *Network) SendFindDataMessage(hash string) {
-	// TODO
-}
-
-func (network *Network) SendStoreMessage(data []byte) {
-	// TODO
-}
