+++ src/kademlia/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"time"
 )
 
@@ -10,14 +9,16 @@ import (
 func main(){
 	for {
 		time.Sleep(time.Second)
-		fmt.Println("Enter 0 for client and 1 for server")
+		/*fmt.Println("Enter 0 for client and 1 for server")
 		var input string
 		fmt.Scanln(&input)
 		if input == "0" {
 			client()
 		}else{
 			server()
-		}
+		}*/
+		node := NewKademliaNode()
+		node.PrintIP()
 	}
 
 }
