+++ src/kademlia/network.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"net"
+	"strings"
+)
+
+type Network struct {
+}
+
+func Listen(ip string, port int) {
+	// TODO
+}
+
+func (network *Network) SendPingMessage(contact *Contact) {
+	// TODO
+}
+
+func (network *Network) SendFindContactMessage(contact *Contact) {
+	// TODO
+}
+
+func (network *Network) SendFindDataMessage(hash string) {
+	// TODO
+}
+
+func (network *Network) SendStoreMessage(data []byte) {
+	// TODO
+}
+
+func GetIP() string{
+	ip := "0.0.0.0"
+	interfaces, _ := net.Interfaces()
+	for _,i  := range interfaces{
+		byNameInterface, _ := net.InterfaceByName(i.Name)
+		if i.Name == "eth0" {
+			addresses, _ := byNameInterface.Addrs()
+			for _, v := range addresses{
+				ip = strings.TrimSuffix(v.String(), "/24")
+			}
+		}
+
+	}
+	return ip
+}
